Level Lock
I dont know how this code, all this is important is that works. After uploading I will retain 0% of how I coded this. ALl you need to know is that it works, and there is no need to fix it. It may look bad, and probably is bad. But it works

The levele variable found in the ENDTRIGGER corresponds to the Level Allowed script which is in the game manager

The next level variable found in the ENDTRIGGER is the next level number after the current level so that level can be unlocked after completion.

The Menu got a new script

The Script changed for the level choose button, now you need to enter the INT of the level to get that level.